+++ lib/event/event.rb
@@ -97,7 +97,7 @@ module StarkBank
     # - user [Project object]: Project object. Not necessary if StarkBank.user was set before function call
     #
     # ## Return:
-    # - deleted Event with updated attributes
+    # - deleted Event object
     def self.delete(id, user: nil)
       StarkBank::Utils::Rest.delete_id(id: id, user: user, **resource)
     end