+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- DasPayment resource
+- IssPayment resource
 
 ## [0.4.2] - 2020-06-24
 ### Changed